Fix mapping type
Change type 'integer' to 'long' for `drug_use.off_label_use.snomed_concept_id`

+++ src/hub/dataload/sources/drugcentral/drugcentral_upload.py
@@ -224,7 +224,7 @@ class DrugCentralUploader(BaseDrugUploader):
                             "type": "keyword"
                         },
                         "snomed_concept_id": {
-                            "type": "integer"
+                            "type": "long"
                         },
                         "concept_name": {
                             "type": "text"
